+++ src/int/mod.rs
@@ -1,9 +1,7 @@
 pub mod area;
 pub mod count;
-pub mod dedup;
 pub mod despike;
 pub mod path;
-pub mod reverse;
 pub mod safe_range;
 pub mod shape;
 pub mod simple;